add specific input rules
on my own system, the idProduct was different from the value expected
by the .rules file, leading to the microphone not being detected.

Furthermore, the generic analog-input-mic was not being mapped correctly
by my pulseaudio. Copying the steelseries-arctis-7 input rule seems to
be detected correctly and functions as expected, however.